+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 
   - Major/Minor/Patch: This is an example changelog item, usually can be the commit message.
   - Append new items to make git merging easier.
+  - Major: Create new `Number` component and integrate it with `CustomFieldInputNumber`
 </details>
 
 ## v0.51.3